Trait sawtooth_sdk::consensus::service::Service

pub trait Service {
    fn send_to(
        &mut self, 
        peer: &PeerId, 
        message_type: &str, 
        payload: Vec<u8>
    ) -> Result<(), Error>;

    fn broadcast(
        &mut self, 
        message_type: &str, 
        payload: Vec<u8>
    ) -> Result<(), Error>;

    fn initialize_block(
        &mut self, 
        previous_id: Option<BlockId>
    ) -> Result<(), Error>;

    fn summarize_block(&mut self) -> Result<Vec<u8>, Error>;

    fn finalize_block(&mut self, data: Vec<u8>) -> Result<BlockId, Error>;

    fn cancel_block(&mut self) -> Result<(), Error>;

    fn check_blocks(&mut self, priority: Vec<BlockId>) -> Result<(), Error>;

    fn commit_block(&mut self, block_id: BlockId) -> Result<(), Error>;

    fn ignore_block(&mut self, block_id: BlockId) -> Result<(), Error>;

    fn fail_block(&mut self, block_id: BlockId) -> Result<(), Error>;

    fn get_blocks(
        &mut self, 
        block_ids: Vec<BlockId>
    ) -> Result<HashMap<BlockId, Block>, Error>;

    fn get_chain_head(&mut self) -> Result<Block, Error>;

    fn get_settings(
        &mut self, 
        block_id: BlockId, 
        keys: Vec<String>
    ) -> Result<HashMap<String, String>, Error>;

    fn get_state(
        &mut self, 
        block_id: BlockId, 
        addresses: Vec<String>
    ) -> Result<HashMap<String, Vec<u8>>, Error>;
}

Provides methods that allow the consensus engine to issue commands and requests.

Required methods

fn send_to(
&mut self,
peer: &PeerId,
message_type: &str,
payload: Vec<u8>
) -> Result<(), Error>

Send a consensus message to a specific, connected peer

fn broadcast(
&mut self,
message_type: &str,
payload: Vec<u8>
) -> Result<(), Error>

Broadcast a message to all connected peers

fn initialize_block(
&mut self,
previous_id: Option<BlockId>
) -> Result<(), Error>

Initialize a new block built on the block with the given previous id and begin adding batches to it. If no previous id is specified, the current head will be used.

fn summarize_block(&mut self) -> Result<Vec<u8>, Error>

Stop adding batches to the current block and return a summary of its contents.

fn finalize_block(&mut self, data: Vec<u8>) -> Result<BlockId, Error>

Insert the given consensus data into the block and sign it. If this call is successful, the consensus engine will receive the block afterwards.

fn cancel_block(&mut self) -> Result<(), Error>

Stop adding batches to the current block and abandon it.

fn check_blocks(&mut self, priority: Vec<BlockId>) -> Result<(), Error>

Update the prioritization of blocks to check

fn commit_block(&mut self, block_id: BlockId) -> Result<(), Error>

Update the block that should be committed

fn ignore_block(&mut self, block_id: BlockId) -> Result<(), Error>

Signal that this block is no longer being committed

fn fail_block(&mut self, block_id: BlockId) -> Result<(), Error>

Mark this block as invalid from the perspective of consensus

fn get_blocks(
&mut self,
block_ids: Vec<BlockId>
) -> Result<HashMap<BlockId, Block>, Error>

Retrieve consensus-related information about blocks

fn get_chain_head(&mut self) -> Result<Block, Error>

Get the chain head block.

fn get_settings(
&mut self,
block_id: BlockId,
keys: Vec<String>
) -> Result<HashMap<String, String>, Error>

Read the value of settings as of the given block

fn get_state(
&mut self,
block_id: BlockId,
addresses: Vec<String>
) -> Result<HashMap<String, Vec<u8>>, Error>

Read values in state as of the given block

Loading content...