Trait sawtooth_sdk::signing::Context

pub trait Context {
    fn get_algorithm_name(&self) -> &str;

    fn sign(
        &self, 
        message: &[u8], 
        key: &dyn PrivateKey
    ) -> Result<String, Error>;

    fn verify(
        &self, 
        signature: &str, 
        message: &[u8], 
        key: &dyn PublicKey
    ) -> Result<bool, Error>;

    fn get_public_key(
        &self, 
        private_key: &dyn PrivateKey
    ) -> Result<Box<dyn PublicKey>, Error>;

    fn new_random_private_key(&self) -> Result<Box<dyn PrivateKey>, Error>;
}

A context for a cryptographic signing algorithm.

Required Methods

fn get_algorithm_name(&self) -> &str

Returns the algorithm name.

fn sign(&self, message: &[u8], key: &dyn PrivateKey) -> Result<String, Error>

Sign a message Given a private key for this algorithm, sign the given message bytes and return a hex-encoded string of the resulting signature.

Arguments

message- the message bytes

private_key the private key

Returns

signature - The signature in a hex-encoded string

fn verify(
&self,
signature: &str,
message: &[u8],
key: &dyn PublicKey
) -> Result<bool, Error>

Verifies that the signature of a message was produced with the associated public key.

Arguments

signature - the hex-encoded signature

message - the message bytes

public_key - the public key to use for verification

Returns

boolean - True if the public key is associated with the signature for that method, False otherwise

fn get_public_key(
&self,
private_key: &dyn PrivateKey
) -> Result<Box<dyn PublicKey>, Error>

Produce the public key for the given private key.

Arguments

private_key - a private key

Returns

public_key - the public key for the given private key

fn new_random_private_key(&self) -> Result<Box<dyn PrivateKey>, Error>

Generates a new random PrivateKey using this context.

Returns

private_key - a random private key